fix: map markers don't appear after the map is shown again
To know if markers are placed into the map, we store that info
into a variable to prevent recreating it (1 marker = 1 asset).

However there is one edge case where the function to create a marker
is triggered while the map isn't ready yet.

The variable is stored but the marker isn't placed and that will
prevent the marker to be placed even if the function is executed again.

Solution: set the map to null when it's hidden